requirements.txt
python-telegram-bot==12.3.0
configparser==4.0.2